pyproject.toml
[project]
name = "capstone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.3.0",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
    "transformers>=4.56.2",
]
